code beautification

+++ CacheSync/Form1.cs
@@ -1,6 +1,7 @@
 ﻿using System;
 using System.IO;
 using System.Windows.Forms;
+
 using ClearCache.Properties;
 
 namespace ClearCache
@@ -9,9 +10,9 @@ namespace ClearCache
     {
         private string _formFolder = "";
         private string _cacheFolder = "";
-        readonly private string _trackFiles;
-        readonly private string[] _clearFiles;
-        readonly private string _cacheFolderName;
+        private readonly string _trackFiles;
+        private readonly string[] _clearFiles;
+        private readonly string _cacheFolderName;
         private FileSystemWatcher _watcher;
 
         public Form1()
@@ -19,10 +20,10 @@ namespace ClearCache
             InitializeComponent();
             textBox_formFolder.Text = _formFolder = Settings.Default.FormFolder;
             textBox_cacheFolder.Text = _cacheFolder = Settings.Default.CacheFolder;
-            _trackFiles = Settings.Default.TrackFiles;
-            _clearFiles = Settings.Default.ClearFiles.Split(';', StringSplitOptions.RemoveEmptyEntries);
+            _trackFiles = Settings.Default.TrackFileMask;
+            _clearFiles = Settings.Default.ClearFileMaskList.Split(';', StringSplitOptions.RemoveEmptyEntries);
             _cacheFolderName = Settings.Default.CacheFolderName;
-            CheckAllDirectoryExists();
+            checkBox_enabled.Enabled = CheckAllDirectoryExists();
         }
 
         private void Button_formFolder_Click(object sender, EventArgs e)
@@ -42,7 +43,7 @@ namespace ClearCache
                 }
             }
 
-            CheckAllDirectoryExists();
+            checkBox_enabled.Enabled = CheckAllDirectoryExists();
         }
 
         private void CheckBox_enabled_CheckedChanged(object sender, EventArgs e)
@@ -50,7 +51,7 @@ namespace ClearCache
             var en = checkBox_enabled.Checked;
             WatcherStart(en);
             button_formFolder.Enabled = !en;
-            textBox_formFolder.Enabled = !en;
+            textBox_formFolder.ReadOnly = en;
         }
 
         private bool WatcherStart(bool en)
@@ -69,12 +70,16 @@ namespace ClearCache
             notifyIcon1.Text = "Active";
             try
             {
-                _watcher = new FileSystemWatcher(_formFolder);
-
-                _watcher.NotifyFilter = NotifyFilters.DirectoryName
+                _watcher = new FileSystemWatcher(_formFolder) {
+                    NotifyFilter = NotifyFilters.DirectoryName
                                         | NotifyFilters.FileName
                                         | NotifyFilters.LastWrite
-                                        | NotifyFilters.Size;
+                                        | NotifyFilters.Size,
+                    Filter = _trackFiles,
+                    IncludeSubdirectories = true,
+                    EnableRaisingEvents = true,
+                    InternalBufferSize = 102400
+                };
 
                 _watcher.Changed += ClearCache;
                 _watcher.Created += ClearCache;
@@ -82,10 +87,6 @@ namespace ClearCache
                 _watcher.Renamed += ClearCache;
                 _watcher.Error += WatcherError;
 
-                _watcher.Filter = _trackFiles;
-                _watcher.IncludeSubdirectories = true;
-                _watcher.EnableRaisingEvents = true;
-                _watcher.InternalBufferSize *= 10;
             }
             catch (Exception ex)
             {
@@ -110,18 +111,12 @@ namespace ClearCache
                 textBox_cacheFolder.Text = _cacheFolder;
             }
 
-            CheckAllDirectoryExists();
+            checkBox_enabled.Enabled = CheckAllDirectoryExists();
         }
-
 
         private bool CheckAllDirectoryExists()
         {
             var result = Directory.Exists(_formFolder); // && Directory.Exists(_cacheFolder);
-            checkBox_enabled.Enabled = result;
-            if (!result)
-            {
-                checkBox_enabled.Checked = false;
-            }
 
             return result;
         }
@@ -130,11 +125,14 @@ namespace ClearCache
         {
             try
             {
-                foreach (var file in _clearFiles)
+                if (Directory.Exists(_cacheFolder))
                 {
-                    foreach (var f in Directory.EnumerateFiles(_cacheFolder, file))
+                    foreach (var file in _clearFiles)
                     {
-                        File.Delete(f);
+                        foreach (var f in Directory.EnumerateFiles(_cacheFolder, file))
+                        {
+                            File.Delete(f);
+                        }
                     }
                 }
             }
